pyproject.toml
[project]
name = "hackapizza_gas"
version = "0.1.0"
description = "Hackapizza solution"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aisuite[groq,openai]>=0.1.8",
    "docling>=2.20.0",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
]
